UI Update, Including Blinker.
I need to tweak a little more and we need to add a returnindex field into SimonCore

I realized since I could not return a value directly with the Action Event item I was going to have to create custom objects to be JButtons. I need to test this on my home system before I say it is completed.